🐛 [#148] 프로필 이미지 출력을 위한 수정

+++ backend/src/main/java/com/Devim/backend/domain/comment/Comment.java
@@ -13,7 +13,7 @@ public class Comment {
 	private long commentNo;
 	private long boardNo;
 	private String commentContent;
-	private int writer;
+	private long writer;
 	private LocalDateTime createdDt;
 	private boolean deleteFlag;
 }
